+++ app/policies/result_policy.rb
@@ -5,10 +5,11 @@ class ResultPolicy < ApplicationPolicy
   alias_rule :create?, :add_extra_mark?, :remove_extra_mark?, :get_test_runs_instructors?,
              :add_tag?, :remove_tag?, :revert_to_automatic_deductions?, :random_incomplete_submission?, to: :grade?
   alias_rule :show?, :get_annotations?, :print?, to: :view?
-  alias_rule :edit?, :update_mark?, :toggle_marking_state?, :update_overall_comment?, :next_grouping?, to: :review?
+  alias_rule :edit?, :toggle_marking_state?, :update_overall_comment?, :next_grouping?, to: :review?
   alias_rule :refresh_view_tokens?, :update_view_token_expiry?, to: :set_released_to_students?
 
   authorize :view_token, optional: true
+  authorize :criterion_id, optional: true
 
   def view?
     check?(:manage_submissions?, role) || check?(:assigned_grader?, record.grouping) ||
@@ -47,6 +48,15 @@ class ResultPolicy < ApplicationPolicy
       record&.submission&.assignment&.has_peer_review &&
           role.is_reviewer_for?(record&.submission&.assignment&.pr_assignment, record)
     )
+  end
+
+  def update_mark?
+    assignment = record.grouping.assignment
+    if !check?(:manage_submissions?, role) && assignment.assign_graders_to_criteria && role.ta?
+      check?(:review?) && check?(:assigned_to_criterion?, with: TaPolicy)
+    else
+      check?(:review?)
+    end
   end
 
   def set_released_to_students?