+++ Changelog.md
@@ -35,6 +35,7 @@
 - Validate copy and page number when fixing errors with scanned exams (#6695)
 - Display due date and submission time in result "Submission Info" tab (#6698)
 - Added "Help" link to the MarkUs page header dropdown (#6702)
+- Prevent TAs that aren't assigned to grade a given criterion from assigning a grade to that criterion (#6699)
 
 ## [v2.2.3]
 - Fix bug where in some circumstances the wrong result would be displayed to students (#6465)